minimizing space used by friendListener
I know it's in bad style to return early, but in this situation it saves
a bunch of time. the friend left -- there is no need to update his
information in the map, and furthermore there is no need to keep his
information as it will be overwritten upon his next login.